requirements.txt
sounddevice
numpy
scipy
openai
python-dotenv
PyQt5
pynput
pyautogui
